WIP: Refactor routes controllers
